空中から着地までの角度の取得

+++ Assets/scripts/Score.cs
@@ -6,7 +6,7 @@ public static class KillScore
 {
     public static int killScore = 0;
     public static int Level = 0;
-    public static int rotateLevel;
+    public static float rotateLevel;
     public static int damageLevel;
     public static int jumpLevel;
     public static bool barrier;
@@ -16,7 +16,7 @@ public static void Reset()
     {
         killScore = 0;
         Level = 0;
-        rotateLevel = 1;
+        rotateLevel = 1f;
         damageLevel = 1;
         jumpLevel = 1;
         barrier = false;
